requirements.txt
requests
psycopg2-binary
beautifulsoup4
currency_converter